+++ src/Smartstore.Core/Catalog/Attributes/Services/ProductAttributeMaterializer.cs
@@ -32,7 +32,6 @@ namespace Smartstore.Core.Catalog.Attributes
         private readonly IRequestCache _requestCache;
         private readonly ICacheManager _cache;
         private readonly Lazy<IDownloadService> _downloadService;
-        private readonly Lazy<CatalogSettings> _catalogSettings;
         private readonly PerformanceSettings _performanceSettings;
 
         public ProductAttributeMaterializer(
@@ -41,7 +40,6 @@ namespace Smartstore.Core.Catalog.Attributes
             IRequestCache requestCache,
             ICacheManager cache,
             Lazy<IDownloadService> downloadService,
-            Lazy<CatalogSettings> catalogSettings,
             PerformanceSettings performanceSettings)
         {
             _db = db;
@@ -49,7 +47,6 @@ namespace Smartstore.Core.Catalog.Attributes
             _requestCache = requestCache;
             _cache = cache;
             _downloadService = downloadService;
-            _catalogSettings = catalogSettings;
             _performanceSettings = performanceSettings;
         }