+++ src/Smartstore.Core/Catalog/Products/Services/ProductCloner.cs
@@ -376,6 +376,7 @@ namespace Smartstore.Core.Catalog.Products
                     ManufacturerPartNumber = combination.ManufacturerPartNumber,
                     Price = combination.Price,
                     AssignedMediaFileIds = combination.AssignedMediaFileIds,
+                    Weight = combination.Weight,
                     Length = combination.Length,
                     Width = combination.Width,
                     Height = combination.Height,