+++ src/Smartstore.Core/Checkout/Shipping/Services/ShippingService.cs
@@ -76,7 +76,7 @@ namespace Smartstore.Core.Checkout.Shipping
 
                     _settingFactory.SaveSettingsAsync(_shippingSettings).GetAwaiter().GetResult();
 
-                    return new Provider<IShippingRateComputationMethod>[] { fallbackProvider };
+                    return [fallbackProvider];
                 }
 
                 if (DataSettings.DatabaseIsInstalled())
@@ -132,19 +132,24 @@ namespace Smartstore.Core.Checkout.Shipping
         {
             Guard.NotNull(cart);
 
-            var cartWeight = await GetCartWeight(cart.Items, true, includeFreeShippingProducts);
-
-            // Checkout attributes.
-            if (cart.Customer != null)
+            var cacheKey = $"shipping-cart-total-weight:{cart.GetHashCode()}-{includeFreeShippingProducts}";
+            var cartWeight = await _requestCache.GetAsync(cacheKey, async () =>
             {
-                var checkoutAttributes = cart.Customer.GenericAttributes.CheckoutAttributes;
-                if (checkoutAttributes.AttributesMap.Any())
-                {
-                    var attributeValues = await _checkoutAttributeMaterializer.MaterializeCheckoutAttributeValuesAsync(checkoutAttributes);
+                var weight = await GetCartWeight(cart.Items, true, includeFreeShippingProducts);
 
-                    cartWeight += attributeValues.Sum(x => x.WeightAdjustment);
+                // Checkout attributes.
+                if (cart.Customer != null)
+                {
+                    var checkoutAttributes = cart.Customer.GenericAttributes.CheckoutAttributes;
+                    if (checkoutAttributes.AttributesMap.Any())
+                    {
+                        var attributeValues = await _checkoutAttributeMaterializer.MaterializeCheckoutAttributeValuesAsync(checkoutAttributes);
+                        weight += attributeValues.Sum(x => x.WeightAdjustment);
+                    }
                 }
-            }
+
+                return weight;
+            });
 
             return cartWeight;
         }
@@ -255,7 +260,7 @@ namespace Smartstore.Core.Checkout.Shipping
             }
             else if (cart.Length == 1)
             {
-                if (IgnoreCartItem(cart[0]))
+                if (IgnoreCartItem(cart[0].Item))
                 {
                     return decimal.Zero;
                 }
@@ -265,11 +270,11 @@ namespace Smartstore.Core.Checkout.Shipping
             else
             {
                 selection = new ProductVariantAttributeSelection(string.Empty);
-                foreach (var item in cart)
+                foreach (var item in cart.Select(x => x.Item))
                 {
                     if (!IgnoreCartItem(item))
                     {
-                        foreach (var attribute in item.Item.AttributeSelection.AttributesMap)
+                        foreach (var attribute in item.AttributeSelection.AttributesMap)
                         {
                             if (!attribute.Value.IsNullOrEmpty())
                             {
@@ -295,17 +300,22 @@ namespace Smartstore.Core.Checkout.Shipping
                 : [];
 
             // INFO: always iterate cart items (not attribute values). Attributes can occur multiple times in cart.
-            foreach (var item in cart)
+            foreach (var item in cart.Select(x => x.Item))
             {
                 if (IgnoreCartItem(item))
                 {
                     continue;
                 }
 
-                var itemWeight = item.Item.Product.Weight;
+                if (item.AttributeSelection.HasAttributes)
+                {
+                    await _productAttributeMaterializer.MergeWithCombinationAsync(item.Product, item.AttributeSelection, null);
+                }
+
+                var itemWeight = item.Product.Weight;
 
                 // Add attributes weight.
-                foreach (var pair in item.Item.AttributeSelection.AttributesMap)
+                foreach (var pair in item.AttributeSelection.AttributesMap)
                 {
                     var integerValues = pair.Value
                         .Select(x => x.ToString())
@@ -335,7 +345,7 @@ namespace Smartstore.Core.Checkout.Shipping
 
                 if (multiplyByQuantity)
                 {
-                    itemWeight *= item.Item.Quantity;
+                    itemWeight *= item.Quantity;
                 }
 
                 cartWeight += itemWeight;
@@ -343,9 +353,9 @@ namespace Smartstore.Core.Checkout.Shipping
 
             return cartWeight;
 
-            bool IgnoreCartItem(OrganizedShoppingCartItem cartItem)
+            bool IgnoreCartItem(ShoppingCartItem cartItem)
             {
-                return !includeFreeShippingProducts && cartItem.Item.Product.IsFreeShipping;
+                return !includeFreeShippingProducts && cartItem.Product.IsFreeShipping;
             }
 
             static string CreateAttributeKey(int productVariantAttributeId, int productVariantAttributeValueId)