+++ src/Smartstore.Core/Platform/DataExchange/Export/ExportXmlHelper.cs
@@ -1027,7 +1027,7 @@ namespace Smartstore.Core.DataExchange.Export
                     _writer.WriteElementString(nameof(ProductVariantAttributeCombination.ManufacturerPartNumber), entityPvac.ManufacturerPartNumber);
 
                     _writer.WriteElementString(nameof(ProductVariantAttributeCombination.Price), entityPvac.Price?.ToString(_culture) ?? string.Empty);
-                    _writer.WriteElementString(nameof(ProductVariantAttributeCombination.Price), entityPvac.Price?.ToString(_culture) ?? string.Empty);
+                    _writer.WriteElementString(nameof(ProductVariantAttributeCombination.Weight), entityPvac.Weight?.ToString(_culture) ?? string.Empty);
                     _writer.WriteElementString(nameof(ProductVariantAttributeCombination.Length), entityPvac.Length?.ToString(_culture) ?? string.Empty);
                     _writer.WriteElementString(nameof(ProductVariantAttributeCombination.Width), entityPvac.Width?.ToString(_culture) ?? string.Empty);
                     _writer.WriteElementString(nameof(ProductVariantAttributeCombination.Height), entityPvac.Height?.ToString(_culture) ?? string.Empty);